Added mudata handling
For MuData inputs, can either pass in the prot data extracted from mudata, then add back into the mudata, or pass in the whole mudata object. If doing the latter, it will look for 'prot' or 'protein' as the modality names, and assume that the desired data to denoise is in .X -- this may need to be improved later for specific layer selection.